Migrate hw-reset/test-sanity.py to pytest
